% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recm.R
\name{recm}
\alias{recm}
\title{Bayesian Radiocarbon-dated Event Count Model.}
\usage{
recm(
  dates,
  X,
  t_edges,
  startvals = NULL,
  niter = 10000,
  adapt = T,
  adapt_amount = 0.1,
  adapt_interval = 100,
  adapt_window = c(0.21, 0.25),
  scales = NULL,
  priors = NULL,
  BP = T,
  calcurve = NULL
)
}
\arguments{
\item{dates}{Matrix containing radiocarbon date means and errors in the
first and second columns, respectively.}

\item{X}{A matrix containing the covariates for the model (independent
variables) including an intercept if desired. Each column should contain
one variable with the leading column being the intercept (a column of 1's).}

\item{startvals}{A numeric vector of starting values for the MCMC. The
leading n elements will be the starting values for the regression
parameters (coefficients associated with the columns of X) while the
remaining columns should contain initial values for the dates (in years BP
1950).}

\item{niter}{Integer indicating the number of MCMC iterations.}

\item{adapt}{Logical (default = True), indicating whether to use an adaptive
MCMC algorithm for finding optimal variances for proposal distributions.}

\item{adapt_amount}{A scalar containing the ratio by which the proposal
variance is changed during the MCMC adapt step when adapt = True---e.g.,
adapt_amount = 0.1 (default) grows/shrinks the variance by 10\%.}

\item{adapt_interval}{(default = 100) A numeric scalar indicating the the n-#'  th mcmc iteration during which the simulation will attempt to adapt the
proposal variances.}

\item{adapt_window}{A numeric vector with two elements that define the lower
and upper bounds of the target acceptance value for the MCMC. This is only
used when adapt=T. Default is \link{0.21, 0.25}.}

\item{scales}{A numeric vector of scales for the proposal distributions. The
order is important. Each element refers to the scale of a proposal function
for the model parameters in the following order: (b_1, b_2, ..., b_nX),
where \code{b} refers to a regression coefficient and `nX' is the number of
columns in X. If Null, naive estimates are used.}

\item{priors}{A numeric vector of parameter values for the model priors---the
order is important and corresponds to order of appearance of the prior
density functions in the source code for the \code{prior} function. If Null,
very wide priors are used--e.g., N(0, 1000).}

\item{BP}{Logical (default = True), indicating whether to assume dates
provided are in BP. If so, dates (ages) increase into the past. It is up to
the user to ensure that the X columns are oriented appropriately.}

\item{calcurve}{Matrix containing a calibration curve. First column should
be calibrated ages/dates; second column should contain radiocarbon years or
fraction modern values; third column should contain errors (standard
deviation). If Null, IntCal::ccurve() is called, meaning that the latest
IntCal curve available from that package is used.}

\item{model}{A string specifying the type of model to use. Currently only
Poisson ('pois') models are supported.}
}
\value{
If adapt = True, returns a named list including: the mcmc
$samples; $acceptance rates for all model parameters; and $scales found
with the adaptive algorithm. Otherwise, returns only the mcmc samples (mcmc
chains).
}
\description{
This function runs a Metropolis-Hastings MCMC simulation to
estimate the parameters of a count-based (Poisson) regression model
accounting for chronological and other uncertainties in both dependent and
independent variables.
}
